requirements.txt
opencv-python
discord.py
aiohttp
aiofiles